# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "ATOS", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/plan": {
      "get": {
        "tags": ["plan"],
        "summary": "Read Plan",
        "operationId": "plan-read_plan",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TrainingPlanPublic" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      },
      "put": {
        "tags": ["plan"],
        "summary": "Replace Plan",
        "operationId": "plan-replace_plan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TrainingPlanUpdate" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TrainingPlanPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/running-plans/": {
      "get": {
        "tags": ["running-plans"],
        "summary": "Read Plans",
        "operationId": "running-plans-read_plans",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RunningPlansPublic" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      },
      "post": {
        "tags": ["running-plans"],
        "summary": "Create Plan",
        "operationId": "running-plans-create_plan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RunningPlanCreate" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RunningPlanPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/running-plans/{plan_id}": {
      "get": {
        "tags": ["running-plans"],
        "summary": "Read Plan",
        "operationId": "running-plans-read_plan",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Plan Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RunningPlanPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["running-plans"],
        "summary": "Replace Plan",
        "operationId": "running-plans-replace_plan",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Plan Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RunningPlanUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RunningPlanPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["running-plans"],
        "summary": "Delete Plan",
        "operationId": "running-plans-delete_plan",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Plan Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": { "type": "string" },
                  "title": "Response Running-Plans-Delete Plan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/running-plans/{plan_id}/workouts/{workout_id}": {
      "patch": {
        "tags": ["running-plans"],
        "summary": "Update Workout",
        "operationId": "running-plans-update_workout",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Plan Id" }
          },
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workout Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RunningWorkoutUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunningWorkoutPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/running-plans/{plan_id}/workouts/{workout_id}/duplicate": {
      "post": {
        "tags": ["running-plans"],
        "summary": "Duplicate Workout",
        "operationId": "running-plans-duplicate_workout",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Plan Id" }
          },
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workout Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  { "$ref": "#/components/schemas/WorkoutDuplicateIn" },
                  { "type": "null" }
                ],
                "title": "Duplicate In"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunningWorkoutPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunningPlanCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "goal": {
            "anyOf": [
              { "type": "string", "maxLength": 2000 },
              { "type": "null" }
            ],
            "title": "Goal"
          },
          "distance_km": {
            "anyOf": [{ "type": "number", "minimum": 0.0 }, { "type": "null" }],
            "title": "Distance Km"
          },
          "distance_unit": {
            "type": "string",
            "enum": ["km", "mi"],
            "title": "Distance Unit",
            "default": "km"
          },
          "target_time_seconds": {
            "anyOf": [
              { "type": "integer", "minimum": 0.0 },
              { "type": "null" }
            ],
            "title": "Target Time Seconds"
          },
          "target_pace_seconds_per_km": {
            "anyOf": [{ "type": "number", "minimum": 0.0 }, { "type": "null" }],
            "title": "Target Pace Seconds Per Km"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              { "type": "string", "format": "date" },
              { "type": "null" }
            ],
            "title": "End Date"
          },
          "notes": {
            "anyOf": [
              { "type": "string", "maxLength": 4000 },
              { "type": "null" }
            ],
            "title": "Notes"
          },
          "status": {
            "type": "string",
            "enum": ["planned", "active", "completed"],
            "title": "Status",
            "default": "planned"
          },
          "race_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Race Id"
          },
          "phases": {
            "items": { "$ref": "#/components/schemas/PhaseIn" },
            "type": "array",
            "title": "Phases"
          }
        },
        "type": "object",
        "required": ["name", "start_date"],
        "title": "RunningPlanCreate"
      },
      "RunningPlanPublic": {
        "properties": {
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "user_id": { "type": "string", "format": "uuid", "title": "User Id" },
          "race_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Race Id"
          },
          "name": { "type": "string", "title": "Name" },
          "goal": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Goal"
          },
          "distance_km": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Distance Km"
          },
          "distance_unit": { "type": "string", "title": "Distance Unit" },
          "target_time_seconds": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Target Time Seconds"
          },
          "target_pace_seconds_per_km": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Target Pace Seconds Per Km"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              { "type": "string", "format": "date" },
              { "type": "null" }
            ],
            "title": "End Date"
          },
          "status": {
            "type": "string",
            "enum": ["planned", "active", "completed"],
            "title": "Status"
          },
          "notes": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "phases": {
            "items": { "$ref": "#/components/schemas/RunningPhasePublic" },
            "type": "array",
            "title": "Phases"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "race_id",
          "name",
          "goal",
          "distance_km",
          "distance_unit",
          "target_time_seconds",
          "target_pace_seconds_per_km",
          "start_date",
          "end_date",
          "status",
          "notes",
          "created_at",
          "updated_at"
        ],
        "title": "RunningPlanPublic"
      },
...
      "RunningPlanUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "goal": {
            "anyOf": [
              { "type": "string", "maxLength": 2000 },
              { "type": "null" }
            ],
            "title": "Goal"
          },
          "distance_km": {
            "anyOf": [{ "type": "number", "minimum": 0.0 }, { "type": "null" }],
            "title": "Distance Km"
          },
          "distance_unit": {
            "type": "string",
            "enum": ["km", "mi"],
            "title": "Distance Unit",
            "default": "km"
          },
          "target_time_seconds": {
            "anyOf": [
              { "type": "integer", "minimum": 0.0 },
              { "type": "null" }
            ],
            "title": "Target Time Seconds"
          },
          "target_pace_seconds_per_km": {
            "anyOf": [{ "type": "number", "minimum": 0.0 }, { "type": "null" }],
            "title": "Target Pace Seconds Per Km"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              { "type": "string", "format": "date" },
              { "type": "null" }
            ],
            "title": "End Date"
          },
          "notes": {
            "anyOf": [
              { "type": "string", "maxLength": 4000 },
              { "type": "null" }
            ],
            "title": "Notes"
          },
          "status": {
            "type": "string",
            "enum": ["planned", "active", "completed"],
            "title": "Status",
            "default": "planned"
          },
          "race_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Race Id"
          },
          "phases": {
            "items": { "$ref": "#/components/schemas/PhaseIn" },
            "type": "array",
            "title": "Phases"
          }
        },
        "type": "object",
        "required": ["name", "start_date"],
        "title": "RunningPlanUpdate",
        "description": "Reemplazo completo del plan: usa los mismos campos que la creaci\u00f3n."
      },
      "RunningPlansPublic": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/RunningPlanSummaryPublic"
            },
            "type": "array",
            "title": "Data"
          },
          "count": { "type": "integer", "title": "Count" }
        },
        "type": "object",
        "required": ["data", "count"],
        "title": "RunningPlansPublic"
      },
...
      "TrainingPlanPublic": {
        "properties": {
          "days": {
            "items": { "$ref": "#/components/schemas/TrainingPlanDayPublic" },
            "type": "array",
            "title": "Days"
          }
        },
        "type": "object",
        "title": "TrainingPlanPublic"
      },
      "TrainingPlanUpdate": {
        "properties": {
          "days": {
            "items": { "$ref": "#/components/schemas/TrainingPlanDayBase" },
            "type": "array",
            "maxItems": 7,
            "title": "Days"
          }
        },
        "type": "object",
        "title": "TrainingPlanUpdate"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": { "scopes": {}, "tokenUrl": "/api/v1/login/access-token" }
        }
      }
    }
  }
}